MutexFactory moved to Cron component folder
- Mutex is used only with crons
